move to `airflow db clean`
