+++ airflow/cli/cli_parser.py
@@ -405,8 +405,8 @@ ARG_RUN_ID = Arg(("-r", "--run-id"), help="Helps to identify this run")
 ARG_CONF = Arg(('-c', '--conf'), help="JSON string that gets pickled into the DagRun's conf attribute")
 ARG_EXEC_DATE = Arg(("-e", "--exec-date"), help="The execution date of the DAG", type=parsedate)
 
-# maintenance
-ARG_MAINTENANCE_TABLES = Arg(
+# db
+ARG_DB_TABLES = Arg(
     ("-t", "--tables"),
     help=lazy_object_proxy.Proxy(
         lambda: f"Table names to perform maintenance on (use comma-separated list).\n"
@@ -414,7 +414,7 @@ ARG_MAINTENANCE_TABLES = Arg(
     ),
     type=string_list_type,
 )
-ARG_MAINTENANCE_TIMESTAMP = Arg(
+ARG_DB_CLEANUP_TIMESTAMP = Arg(
     ("--clean-before-timestamp",),
     help="The date or timestamp before which data should be purged.\n"
     "If no timezone info is supplied then dates are assumed to be in airflow default timezone.\n"
@@ -422,7 +422,7 @@ ARG_MAINTENANCE_TIMESTAMP = Arg(
     type=parsedate,
     required=True,
 )
-ARG_MAINTENANCE_DRY_RUN = Arg(
+ARG_DB_DRY_RUN = Arg(
     ("--dry-run",),
     help="Perform a dry run",
     action="store_true",
@@ -1107,20 +1107,6 @@ DAGS_COMMANDS = (
         args=(ARG_CLEAR_ONLY,),
     ),
 )
-MAINTENANCE_COMMANDS = (
-    ActionCommand(
-        name='cleanup-tables',
-        help="Purge old records in metastore tables",
-        func=lazy_load_command('airflow.cli.commands.maintenance_command.cleanup_tables'),
-        args=(
-            ARG_MAINTENANCE_TABLES,
-            ARG_MAINTENANCE_DRY_RUN,
-            ARG_MAINTENANCE_TIMESTAMP,
-            ARG_VERBOSE,
-            ARG_YES,
-        ),
-    ),
-)
 TASKS_COMMANDS = (
     ActionCommand(
         name='list',
@@ -1352,6 +1338,18 @@ DB_COMMANDS = (
         help="Check if the database can be reached",
         func=lazy_load_command('airflow.cli.commands.db_command.check'),
         args=(),
+    ),
+    ActionCommand(
+        name='clean',
+        help="Purge old records in metastore tables",
+        func=lazy_load_command('airflow.cli.commands.db_command.cleanup_tables'),
+        args=(
+            ARG_DB_TABLES,
+            ARG_DB_DRY_RUN,
+            ARG_DB_CLEANUP_TIMESTAMP,
+            ARG_VERBOSE,
+            ARG_YES,
+        ),
     ),
 )
 CONNECTIONS_COMMANDS = (
@@ -1678,11 +1676,6 @@ airflow_commands: List[CLICommand] = [
         name='tasks',
         help='Manage tasks',
         subcommands=TASKS_COMMANDS,
-    ),
-    GroupCommand(
-        name='maintenance',
-        help='Run maintenance tasks',
-        subcommands=MAINTENANCE_COMMANDS,
     ),
     GroupCommand(
         name='pools',
