+++ airflow/cli/commands/db_command.py
@@ -22,6 +22,7 @@ from tempfile import NamedTemporaryFile
 from airflow import settings
 from airflow.exceptions import AirflowException
 from airflow.utils import cli as cli_utils, db
+from airflow.utils.metastore_cleanup import config_dict, run_cleanup
 from airflow.utils.process_utils import execute_interactive
 
 
@@ -101,3 +102,26 @@ def shell(args):
 def check(_):
     """Runs a check command that checks if db is available."""
     db.check()
+
+
+@cli_utils.action_cli(check_db=False)
+def clean(args):
+    """Upgrades the metadata database"""
+    print("DB: " + repr(settings.engine.url))
+    db.upgradedb()
+
+
+# lazily imported by CLI parser for `help` command
+all_tables = list(sorted(config_dict.keys()))
+
+
+@cli_utils.action_cli(check_db=False)
+def cleanup_tables(args):
+    """Purges old records in metastore database"""
+    run_cleanup(
+        table_names=args.tables,
+        dry_run=args.dry_run,
+        clean_before_timestamp=args.clean_before_timestamp,
+        verbose=args.verbose,
+        confirm=not args.yes,
+    )
